Reduce repetition by extracting backend implementation into a macro.
